@@ -32,40 +32,69 @@
             <name>Sean Scanlon</name>
             <email>sean.scanlon@gmail.com</email>
         </developer>
+        <developer>
+            <id>flozano</id>
+            <name>Francisco A. Lozano</name>
+            <email>flozano@gmail.com</email>
+            <roles><role>version updates</role></roles>
+        </developer>
     </developers>
     <properties>
-        <org.springframework.version>4.0.3.RELEASE</org.springframework.version>
-        <jmustache.version>1.9</jmustache.version>
-        <mustache.java.version>0.8.14</mustache.java.version>
-        <mockito.version>1.9.0</mockito.version>
-        <junit.version>4.11</junit.version>
-        <jdk.version>1.5</jdk.version>
+        <org.springframework.version>6.0.0</org.springframework.version>
+        <jmustache.version>1.15</jmustache.version>
+        <mustache.java.version>0.9.10</mustache.java.version>
+        <mockito.version>4.9.0</mockito.version>
+        <junit.version>4.13.2</junit.version>
+        <servlet-api.version>6.0.0</servlet-api.version>
     </properties>
-    <profiles>
-	    <profile>
-		    <id>disable-java8-doclint</id>
-		    <activation>
-			    <jdk>[1.8,)</jdk>
-		    </activation>
-		    <properties>
-			    <additionalparam>-Xdoclint:none</additionalparam>
-		    </properties>
-	    </profile>
-    </profiles>
     <build>
         <plugins>
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>2.3.2</version>
+                <version>3.10.1</version>
                 <configuration>
-                    <source>${jdk.version}</source>
-                    <target>${jdk.version}</target>
+                    <release>17</release>
                 </configuration>
+            </plugin>
+            <plugin>
+                <artifactId>maven-source-plugin</artifactId>
+                <version>3.2.1</version>
+                <executions>
+                    <execution>
+                        <id>attach-sources</id>
+                        <goals>
+                            <goal>jar-no-fork</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <artifactId>maven-javadoc-plugin</artifactId>
+                <version>3.4.1</version>
+                <executions>
+                    <execution>
+                        <id>attach-javadocs</id>
+                        <goals>
+                            <goal>jar</goal>
+                        </goals>
+                        <configuration>
+                            <doclint>none</doclint>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <artifactId>maven-site-plugin</artifactId>
+                <version>3.12.1</version>
+            </plugin>
+            <plugin>
+                <artifactId>maven-project-info-reports-plugin</artifactId>
+                <version>3.4.1</version>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-enforcer-plugin</artifactId>
-                <version>1.2</version>
+                <version>3.1.0</version>
                 <executions>
                     <execution>
                         <id>enforce</id>
