@@ -15,9 +15,7 @@
  */
 package org.springframework.web.servlet.view.mustache;
 
-/**
- * @author Sean Scanlon <sean.scanlon@gmail.com>
- */
+
 public interface MustacheTemplateFactory {
 
     MustacheTemplate getTemplate(String templateURL) throws MustacheTemplateException;
