@@ -1,15 +1,16 @@
 package org.springframework.web.servlet.view.mustache.java;
 
-import com.github.mustachejava.Mustache;
+import static org.mockito.Mockito.verify;
+
+import java.io.Writer;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
-import java.io.Writer;
-
-import static org.mockito.Mockito.verify;
+import com.github.mustachejava.Mustache;
 
 @RunWith(MockitoJUnitRunner.class)
 public class MustacheJTemplateTest {
