@@ -15,16 +15,17 @@
  */
 package org.springframework.web.servlet.view.mustache.jmustache;
 
-import com.samskivert.mustache.Template;
+import static org.mockito.Mockito.verify;
+
+import java.io.Writer;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
-import java.io.Writer;
-
-import static org.mockito.Mockito.verify;
+import com.samskivert.mustache.Template;
 
 @RunWith(MockitoJUnitRunner.class)
 public class JMustacheTemplateTest {
